[FEATURE] Enable readable anchors to TYPO3 content elements
In this current state, the new field uses the TCA type 'slug'.
This allows to generate anchor slugs from the current header,
as well as evaluate an unique slug for all content elements
on the same page.

The downside of the current approach is that the slug field
becomes mandatory. The editor cannot save a content element
without a slug being set.