[TASK] Rename 'anchor' to 'fragment'
'Fragment' (or 'fragment identifier') is the official name in HTML.
'Anchor' is a commonly used term for this part of a URL though.

We'll use the official name for technical parts of the extension.
Editor-related labels will continue to use 'anchor'.

This also allows us to introduce a new TCA checkbox, which an editor
can use to set anchor links next to headers.