[FEATURE] Simplify fragment setup with TypoScript
If an editor sets a fragment identifier to a content element which
matches the ID of a template-related HTML element, this could
result in breaking the website:
The ID attribute could be styled with CSS or accessed with JavaScript.

By adding a prefix or suffix to the fragment identifier,
we can minimize the risk.

Until now, a UID prefix had to be maintained multiple times in
the Fluid templates:
- "c{uid}-{fragmentIdentifier}')}" in the header partials;
- "c{element.data.uid}-{element.data.tx_content_slug_fragment}" in the
  Section Index templates.
This was cumbersome and prone to error.

With the new TypoScript setup, we can now compose our complete
fragment identifier once and then render "{fragmentIdentifier}" or
"{element.fragmentIdentifier}" in the templates.

By default, the UID is still used as a prefix. The TypoScript setup
allows customizing prefix and/or suffix as needed.

The new DataProcessor "FragmentIdentifierProcessor" will process
the fragment for the Section Index menus and checks for hidden headers
(which cannot be linked).

Fixes #6