[TASK] Prefix fragment identifier with current uid
As the fragment identifier is used as an id attribute in HTML,
this could match the id of an element in the website's template
(e.g. '#main-navigation').

This template-related identifier could be styled with CSS,
or be accessed with JavaScript.

Most likely, this would end in unexpected side effects.

By adding the current uid, we'll circumvent this potential issue.