[FEATURE] Override fragments in RTE and TCA links
The new 'ReplaceFragment' hook allows overwriting URL fragments in
page links. It uses the 'modifyPageLinkConfiguration' hook from the
TYPO3 Core. The new Typoscript variable 'fragmentIdentifier' is
processed to get the complete URL fragment, including a possible
custom prefix or suffix.

The hook can be disabled with TypoScript.

The stored link in the rich text or TCA field remains unchanged.
This prevents breaking links in case the human-readable fragment is
modified subsequently, or this extension is uninstalled.

Fixes #5