[!!!][TASK] Switch to Symfony dependency injection (DI)
The Extbase ObjectManager is deprecated in TYPO3 v11.
Therefore, this extension now uses method injection to get the
TypoScript configuration.
Symfony based DI was introduced in TYPO3 10.0 (Feature #84112),
which makes this change incompatible with previous TYPO3 versions.

In Services.yaml, 'public: true' must be set for both
the Fluid data processor and the hook.