[!!!][FEATURE] Allow using header field as fragment source
It is now possible to use the header as a fallback or even
the primary source for human-readable slugs.

This might be a breaking change if you removed the 'if' condition
of the default TypoScript setup.

To enable this, extend/adjust the default TypoScript like this:

lib.contentElement.variables.fragmentIdentifier {
  // Don't check if the fragment field is filled:
  if >
  // If the fragment field is empty, use the header field as a fallback:
  20.ifEmpty.field = header
  // Sanitize the value of the header field:
  stdWrap.postUserFunc = Sebkln\ContentSlug\Evaluation\FragmentEvaluation->sanitizeFragment
}

Resolves #4