[BUGFIX] Fix PHP 8.1 warning in hook and event listener
In a TYPO3 installation with multiple page trees,
this extension might not be configured for every website.

The necessary TypoScript setup is then missing when linking to
a content element from within the RTE or field with type inputLink,
resulting in a PHP warning.

The null coalescing operator prevents this issue.