[FEATURE] Add option to disable check for hidden headers
By default, links to a content element with hidden header will not
use a given human-readable fragment. This is because the intended
use case of this extension are anchors on heading elements.

The new setting "checkForHiddenHeaders" allows to disable this check.
The TYPO3 administrator or integrator is then responsible to provide
a suitable Fluid layout for content elements that renders
the {fragmentIdentifier} at a different HTML tag.

Fixes #20
Closes #24