[BUGFIX] Generate anchor from header when mass-editing records
The Single Table View in the List module allows to edit multiple records
with selected field at the same time.

This change fixes the button to automatically generate a human-readable
anchor from the current header in this editing view.
Requirement: The header field must be rendered in the view as well.

Fixes #9